[#9317] Separate AgentCollectionService
