+++ web/src/main/java/com/navercorp/pinpoint/web/service/AgentInfoService.java
@@ -17,21 +17,11 @@
 package com.navercorp.pinpoint.web.service;
 
 import com.navercorp.pinpoint.common.server.util.time.Range;
-import com.navercorp.pinpoint.web.vo.tree.AgentsMapByApplication;
-import com.navercorp.pinpoint.web.vo.tree.ApplicationAgentHostList;
-import com.navercorp.pinpoint.web.vo.tree.AgentsMapByHost;
 import com.navercorp.pinpoint.web.vo.agent.AgentAndStatus;
 import com.navercorp.pinpoint.web.vo.agent.AgentInfo;
-import com.navercorp.pinpoint.web.vo.agent.AgentInfoFilter;
 import com.navercorp.pinpoint.web.vo.agent.AgentStatus;
-import com.navercorp.pinpoint.web.vo.agent.AgentStatusQuery;
 import com.navercorp.pinpoint.web.vo.agent.DetailedAgentAndStatus;
 import com.navercorp.pinpoint.web.vo.timeline.inspector.InspectorTimeline;
-import com.navercorp.pinpoint.web.vo.tree.SortByAgentInfo;
-
-import java.util.List;
-import java.util.Optional;
-import java.util.Set;
 
 /**
  * @author netspider
@@ -39,33 +29,15 @@ import java.util.Set;
  */
 public interface AgentInfoService {
 
-    int NO_DURATION = -1;
+    AgentAndStatus getAgentAndStatus(String agentId, long timestamp);
 
-    AgentsMapByApplication getAllAgentsList(AgentInfoFilter filter, long timestamp);
-
-    AgentsMapByHost getAgentsListByApplicationName(AgentInfoFilter filter, String applicationName, long timestamp);
-
-    AgentsMapByHost getAgentsListByApplicationName(AgentInfoFilter filter, String applicationName, long timestamp, SortByAgentInfo.Rules sortBy);
-
-    ApplicationAgentHostList getApplicationAgentHostList(int offset, int limit, int durationDays);
-
-    Set<AgentAndStatus> getAgentsByApplicationName(String applicationName, long timestamp);
-
-    Set<AgentInfo> getAgentsByApplicationNameWithoutStatus(String applicationName, long timestamp);
-
-    Set<AgentAndStatus> getRecentAgentsByApplicationName(String applicationName, long timestamp, long timeDiff);
-
-    AgentAndStatus getAgentInfo(String agentId, long timestamp);
-
-    DetailedAgentAndStatus getDetailedAgentInfo(String agentId, long timestamp);
+    DetailedAgentAndStatus getDetailedAgentAndStatus(String agentId, long timestamp);
 
     AgentInfo getAgentInfoWithoutStatus(String agentId, long timestamp);
 
     AgentInfo getAgentInfoWithoutStatus(String agentId, long agentStartTime, int deltaTimeInMilliseconds);
 
     AgentStatus getAgentStatus(String agentId, long timestamp);
-
-    List<Optional<AgentStatus>> getAgentStatus(AgentStatusQuery query);
 
     boolean isActiveAgent(String agentId, Range range);
